requirements.txt
requests>=2.31.0
jsonschema>=4.17.0
python-dotenv>=1.0.0
slack_sdk>=3.21.0
